refactory: fix import auth service
