+++ src/modules/auth/auth.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
 import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
 import { AuthGuard } from './auth.guard';
@@ -8,6 +7,7 @@ import { APP_GUARD } from '@nestjs/core';
 import { JwtModule } from '@nestjs/jwt';
 import { jwtConstants } from './constants/jwt.constants';
 import { UserModule } from '../user/user.module';
+import { AuthService } from './application/auth.service';
 
 @Module({
   imports: [
